addGroupOfUnusedAnimals adds a group to the saved groups if needed

addGroupOfUnusedAnimals adds a group to the saved groups if needed

Usage

addGroupOfUnusedAnimals(savedGroupMembers, candidates, ped, minAge, harem)

Arguments

savedGroupMembers

list of groups of animals in the form of a vector of animal Ids.

candidates

character vector of IDs of the animals available for use in the group.

ped

dataframe that is the `Pedigree`. It contains pedigree information including the IDs listed in candidates.

minAge

integer value indicating the minimum age to consider in group formation. Pairwise kinships involving an animal of this age or younger will be ignored. Default is 1 year.

harem

logical variable when set to TRUE, the formed groups have a single male at least minAge old.

Value

A list of groups, which are each lists of animal Ids that are unused animals at the end of the iteration.