Form age pyramid plot

Form age pyramid plot

Usage

agePyramidPlot(
  males,
  females,
  ageLabels,
  mcol,
  fcol,
  laxlab,
  raxlab,
  gap,
  currentDate
)

Arguments

males

integer vector with the number of males in age groups corresponding to the position in the vector

females

integer vector with the number of females in age groups corresponding to the position in the vector

ageLabels

character vector of labels for the categories represented by each pair of bars. There should be a label for each lx or rx value, even if empty. If labels is a matrix or data frame, the first two columns will be used for the left and right category labels respectively.

mcol

color for the male (left) side of the plot

fcol

color for the female (right) side of the plot

laxlab

label for the male (left) side of the plot

raxlab

label for the female (right) side of the plot

gap

numeric value for one half of the space between the two sets of bars for the ageLabels in user units

currentDate

POSIXct date object indicating the date corresponding to the date the pedigree census occurred.

Value

The return value of par("mar") when the function was called.