Forms and fills list of animals groups based on provided constraints

Forms and fills list of animals groups based on provided constraints

fillGroupMembers(
  candidates,
  currentGroups,
  kin,
  ped,
  harem,
  minAge,
  numGp,
  sexRatio
)

Arguments

candidates
character vector of IDs of the animals available for use in the group.

currentGroups
list of character vectors of IDs of animals currently assigned to the group. Defaults to character(0) assuming no groups are existent.

kin
list of animals and those animals who are related above a threshold value.

ped
dataframe that is the `Pedigree`. It contains pedigree information including the IDs listed in candidates.

harem
logical variable when set to TRUE, the formed groups have a single male at least minAge old.

minAge
integer value indicating the minimum age to consider in group formation. Pairwise kinships involving an animal of this age or younger will be ignored. Default is 1 year.

numGp
integer value indicating the number of groups that should be formed from the list of IDs. Default is 1.

sexRatio
numeric value indicating the ratio of females to males x (from 0.5 to 20 by increments of 0.5 within the accompanying Shiny application. A sex ratio of 0 ignores sex in making up groups.

Value

A list of animal groups and their member animals