Genetic Management Functions

Primary Data Structure --- Pedigree

Contains studbook information for a number of individuals. ASSUME: All IDs listed in the sire or dam columns must have a row entry in the id column

See also