Get site information

Get site information

getSiteInfo(expectConfigFile = TRUE)

Arguments

expectConfigFile
logical parameter when set to FALSE, no configuration is looked for. Default value is TRUE.

Value

A list of site specific information used by the application.

Currently this returns the following character strings in a named list.

centerOne of "SNPRC" or "ONPRC"

baseUrlIf center is "SNPRC", baseUrl is one of "https://boomer.txbiomed.local:8080/labkey" or "https://vger.txbiomed.local:8080/labkey". To allow testing, if center is "ONPRC" baseUrl is "https://boomer.txbiomed.local:8080/labkey".

schemaNameIf center is "SNPRC", schemaName is "study". If center is "ONPRC", schemaName is "study"

folderPath If center is "SNPRC", folderPath is "/SNPRC". If center is "ONPRC", folderPath is "/ONPRC"

queryNameis "demographics"

Examples

# \donttest{
library(nprcgenekeepr)
getSiteInfo()

#> Warning: The nprcgenekeepr configuration file is missing.
#> It is required when the LabKey API is to be used.
#> The file should be named: ~/.nprcgenekeepr_config.

#> $center
#> [1] "ONPRC"
#> 
#> $baseUrl
#> [1] "https://primeuat.ohsu.edu"
#> 
#> $schemaName
#> [1] "study"
#> 
#> $folderPath
#> [1] "/ONPRC/EHR"
#> 
#> $queryName
#> [1] "demographics"
#> 
#> $lkPedColumns
#> [1] "Id"              "gender"          "birth"           "death"          
#> [5] "lastDayAtCenter" "Id/parents/dam"  "Id/parents/sire"
#> 
#> $mapPedColumns
#> [1] "id"    "sex"   "birth" "death" "exit"  "dam"   "sire" 
#> 
#> $sysname
#> [1] "Darwin"
#> 
#> $release
#> [1] "20.2.0"
#> 
#> $version
#> [1] "Darwin Kernel Version 20.2.0: Wed Dec  2 20:39:59 PST 2020; root:xnu-7195.60.75~1/RELEASE_X86_64"
#> 
#> $nodename
#> [1] "prefect.fios-router.home"
#> 
#> $machine
#> [1] "x86_64"
#> 
#> $login
#> [1] "root"
#> 
#> $user
#> [1] "msharp"
#> 
#> $effective_user
#> [1] "msharp"
#> 
#> $homeDir
#> [1] "~/"
#> 
#> $configFile
#> [1] "~/.nprcgenekeepr_config"
#> 

# }