Add animals to an existing breeding group or forms groups:

groupAddAssign finds the largest group that can be formed by adding unrelated animals from a set of candidate IDs to an existing group, to a new group it has formed from a set of candidate IDs or if more than 1 group is desired, it finds the set of groups with the largest average size.The function implements a maximal independent set (MIS) algorithm to find groups of unrelated animals. A set of animals may have many different MISs of varying sizes, and finding the largest would require traversing all possible combinations of animals. Since this could be very time consuming, this algorithm produces a random sample of the possible MISs, and selects from these. The size of the random sample is determined by the specified number of iterations.

Usage

groupAddAssign(
  candidates,
  kmat,
  ped,
  currentGroups = list(character(0L)),
  threshold = 0.015625,
  ignore = list(c("F", "F")),
  minAge = 1,
  iter = 1000L,
  numGp = 1L,
  harem = FALSE,
  sexRatio = 0,
  withKin = FALSE,
  updateProgress = NULL
)

Arguments

candidates

Character vector of IDs of the animals available for use in forming the groups. The animals that may be present in currentGroups are not included within candidates.

kmat

Numeric matrix of pairwise kinship values. Rows and columns are named with animal IDs.

ped

Dataframe that is the Pedigree. It contains pedigree information including the IDs listed in candidates.

currentGroups

List of character vectors of IDs of animals currently assigned to groups. Defaults to a list with character(0) in each sublist element (one for each group being formed) assuming no groups are prepopulated.

threshold

Numeric value indicating the minimum kinship level to be considered in group formation. Pairwise kinship below this level will be ignored. The default value is 0.015625.

ignore

List of character vectors representing the sex combinations to be ignored. If provided, the vectors in the list specify if pairwise kinship should be ignored between certain sexes. Default is to ignore all pairwise kinship between females.

minAge

Integer value indicating the minimum age to consider in group formation. Pairwise kinships involving an animal of this age or younger will be ignored. Default is 1 year.

iter

Integer indicating the number of times to perform the random group formation process. Default value is 1000 iterations.

numGp

Integer value indicating the number of groups that should be formed from the list of IDs. Default is 1.

harem

Logical variable when set to TRUE, the formed groups have a single male at least minAge old.

sexRatio

Numeric value indicating the ratio of females to males x from 0.5 to 20 by increments of 0.5.

withKin

Logical variable when set to TRUE, the kinship matrix for the group is returned along with the group and score. Defaults to not return the kinship matrix. This maintains compatibility with earlier versions.

updateProgress

Function or NULL. If this function is defined, it will be called during each iteration to update a shiny::Progress object.

Value

A list with list items group, score and optionally groupKin. The list item group contains a list of the best group(s) produced during the simulation. The list item score provides the score associated with the group(s). The list item groupKin contains the subset of the kinship matrix that is specific for each group formed.

Details

Part of Group Formation

Examples

library(nprcgenekeepr)
examplePedigree <- nprcgenekeepr::examplePedigree
breederPed <- qcStudbook(examplePedigree,
  minParentAge = 2,
  reportChanges = FALSE,
  reportErrors = FALSE
)
focalAnimals <- breederPed$id[!(is.na(breederPed$sire) &
  is.na(breederPed$dam)) &
  is.na(breederPed$exit)]
ped <- setPopulation(ped = breederPed, ids = focalAnimals)
trimmedPed <- trimPedigree(focalAnimals, breederPed)
probands <- ped$id[ped$population]
ped <- trimPedigree(probands, ped,
  removeUninformative = FALSE,
  addBackParents = FALSE
)
geneticValue <- reportGV(ped,
  guIter = 50, # should be >= 1000
  guThresh = 3,
  byID = TRUE,
  updateProgress = NULL
)
trimmedGeneticValue <- reportGV(trimmedPed,
  guIter = 50, # should be >= 1000
  guThresh = 3,
  byID = TRUE,
  updateProgress = NULL
)
candidates <- trimmedPed$id[trimmedPed$birth < as.Date("2013-01-01") &
  !is.na(trimmedPed$birth) &
  is.na(trimmedPed$exit)]
haremGrp <- groupAddAssign(
  kmat = trimmedGeneticValue[["kinship"]],
  ped = trimmedPed,
  candidates = candidates,
  iter = 10, # should be >= 1000
  numGp = 6,
  harem = TRUE
)
haremGrp$group
#> [[1]]
#>  [1] "XX0GYV" "YFCIHJ" "5EDLL7" "ZQXZYB" "NK802Y" "LVYYNY" "1SPLS8" "QCA36T"
#>  [9] "RVHVTZ" "7ZNY75" "I5CI33" "1KJ2MG" "2Z4YLY" "TEACA3" "S3EBGZ" "5IAFMK"
#> [17] "9MG040" "967Y3D" "50D77I" "D4B0RM" "5KWNMZ" "BCJJKN" "3YJIMV" "NN3GDQ"
#> [25] "FB5L3N" "N79QXB" "LS184H" "1SSCJC" "CS23RV" "K3TNHP"
#> 
#> [[2]]
#>  [1] "BTTHAJ" "PYPM1W" "S222R3" "WK89I9" "DRXMW4" "465ERA" "JLFKV8" "WLMGS1"
#>  [9] "PVY432" "9P0DES" "30J3CQ" "QRZK48" "0XTZQ1" "3GECJJ" "DPXEQE" "EMV4P6"
#> [17] "Q7U139" "DHNQ1W" "WNEAS6" "13B1QL" "5ERY5Z" "0HYZ23" "S7IWWA" "AIHJ8Z"
#> [25] "B1WVCN" "DCJJYS" "N5QBWD" "7NE2UT" "AZ3L0D" "414N7M" "GAS52W" "W0GUKI"
#> 
#> [[3]]
#>  [1] "S63QDN" "SHG3RB" "GCBYDW" "S056D5" "XYRDKV" "QCENKM" "3DTD2N" "XFWVVX"
#>  [9] "YLRNIK" "MH88T6" "0IIAEN" "G8MCV7" "5W621W" "X694YR" "GTLA8R" "46ZHKN"
#> [17] "B134XZ" "H2J6UA" "J3F6PD" "MKY9TK" "Q17CG3" "TXZUKC" "AFZKBS" "C18V6I"
#> [25] "FL170P" "1GF3GM" "R5AYJK" "3SKITJ" "MX4J7G" "1QVS67" "PJ72W1" "GIIEUD"
#> [33] "SH3FB7"
#> 
#> [[4]]
#>  [1] "EEGLWY" "WJXIH9" "5BPBUI" "Y6DB6L" "ZH3YG1" "G58RGY" "PBAFJF" "Q8U9LB"
#>  [9] "QQMBT1" "WTE53B" "BS3RLE" "W5WIRP" "321LLB" "KEA4QG" "LYSLPP" "D9P18Y"
#> [17] "S5H1GC" "CLSVU6" "DKIM6U" "IH1KPA" "E5Q33K" "83HQBN" "F7I2ED" "AR17R5"
#> [25] "F45799" "99BMJW" "W6MDVK" "87AQLF" "MTCAIG" "RJ4JPC" "N4NV8B" "Z904TJ"
#> 
#> [[5]]
#>  [1] "GDXWJ1" "I8ABC7" "G25E3F" "DH9WJQ" "CRPXY7" "CMMUKU" "AP1YLW" "PI4VHT"
#>  [9] "WKY2SZ" "TQEMY6" "0SGJ12" "HE0SCR" "7RA57Q" "1CIRC9" "7B9CA6" "FJS7RQ"
#> [17] "QW2Z3R" "YTJ2UL" "Y0TCYX" "AW400C" "ESUIAF" "WI38KZ" "2F6J3U" "XEC0M5"
#> [25] "QWKFBH" "6X6BG9" "DI4AHD" "B228Q6" "SCFSBF"
#> 
#> [[6]]
#>  [1] "MEUZ85" "ILVQVB" "CHK1ZX" "ZATMEE" "MYUMMX" "BKWE4D" "ZPS15A" "72LYDE"
#>  [9] "KZY6PD" "MFKT9C" "92UG4N" "1FAZ0K" "D33J06" "38K2SR" "5EDIEE" "T3QPW5"
#> [17] "PU7RSG" "M9PVG5" "0X4W26" "01QRQ4" "8JUUJ9" "EZ2F8A" "EX5K0S" "1CZM30"
#> [25] "1VP3UC" "6KWVRI" "FG0SFA" "TYEWF1" "MB6NYQ" "6F9FB8" "VWC5ZH" "AR5U44"
#> [33] "KX0RJ3" "MPIQ4N"
#> 
#> [[7]]
#> [1] NA
#> 
sexRatioGrp <- groupAddAssign(
  kmat = trimmedGeneticValue[["kinship"]],
  ped = trimmedPed,
  candidates = candidates,
  iter = 10L, # should be >= 1000L
  numGp = 6L,
  sexRatio = 9.0
)
sexRatioGrp$group
#> [[1]]
#>  [1] "D33J06" "ZDRSG0" "IH1KPA" "MYUMMX" "WKY2SZ" "ZH3YG1" "RJ4JPC" "RVHVTZ"
#>  [9] "NK802Y" "YFCIHJ" "QRZK48" "JLFKV8" "FJS7RQ" "YDRD81" "99BMJW" "AFZKBS"
#> [17] "GAS52W" "9P0DES" "XEC0M5" "92UG4N" "CHK1ZX" "PJ72W1" "0SGJ12" "TEACA3"
#> [25] "K7900I"
#> 
#> [[2]]
#>  [1] "DH9WJQ" "3YHBC1" "CRPXY7" "T3QPW5" "D9P18Y" "TXZUKC" "WK89I9" "5EDIEE"
#>  [9] "R5AYJK" "YTJ2UL" "5KWNMZ" "YLRNIK" "K3TNHP" "NHWTJ9" "ESUIAF" "G58RGY"
#> [17] "7RA57Q" "MB6NYQ" "5IAFMK" "B228Q6" "CLSVU6" "46ZHKN" "BS3RLE" "Q17CG3"
#> [25] "A98D7P" "KEA4QG" "TQEMY6"
#> 
#> [[3]]
#>  [1] "1SSCJC" "Z7NBA2" "5BPBUI" "LYSLPP" "MX4J7G" "MTCAIG" "KX0RJ3" "50D77I"
#>  [9] "AR5U44" "1KJ2MG" "1FAZ0K" "6X6BG9" "N5QBWD" "R6HV9A" "LS184H" "MFKT9C"
#> [17] "2Z4YLY" "KZY6PD" "6F9FB8" "GCBYDW" "5ERY5Z" "MPIQ4N" "01QRQ4" "CS23RV"
#> [25] "JSAP3H" "S222R3" "M9PVG5"
#> 
#> [[4]]
#>  [1] "HE0SCR" "B2YJJP" "C18V6I" "I5CI33" "321LLB" "N79QXB" "NN3GDQ" "G8MCV7"
#>  [9] "G25E3F" "WTE53B" "8JUUJ9" "B134XZ" "Q7U139" "A6A1M1" "FG0SFA" "ILVQVB"
#> [17] "F7I2ED" "414N7M" "PU7RSG" "1QVS67" "X694YR" "PBAFJF" "E5Q33K" "AR17R5"
#> [25] "IZDV8K"
#> 
#> [[5]]
#>  [1] "DI4AHD" "XY2CK7" "3GECJJ" "3SKITJ" "DPXEQE" "F45799" "WLMGS1" "QQMBT1"
#>  [9] "7B9CA6" "83HQBN" "1SPLS8" "ZATMEE" "9MG040" "LMJWTN" "QWKFBH" "VWC5ZH"
#> [17] "FL170P" "MKY9TK" "Y6DB6L" "BKWE4D" "LVYYNY" "DCJJYS" "TYEWF1" "S3EBGZ"
#> [25] "Z25D52" "WNEAS6" "D4B0RM" "AW400C" "EZ2F8A" "GIIEUD" "S056D5" "13B1QL"
#> [33] "SHG3RB" "Y0TCYX" "3MMZD4"
#> 
#> [[6]]
#>  [1] "FB5L3N" "5KFB90" "WI38KZ" "6KWVRI" "1CIRC9" "38K2SR" "SH3FB7" "H2J6UA"
#>  [9] "MH88T6" "0X4W26" "PVY432" "SCFSBF" "QCENKM" "KZM9RB" "30J3CQ" "B1WVCN"
#> [17] "3YJIMV" "AP1YLW" "1GF3GM" "PI4VHT" "Q8U9LB" "XYRDKV" "72LYDE" "7NE2UT"
#> [25] "G2GYST"
#> 
#> [[7]]
#>   [1] "HLQ9SY" "XFWVVX" "B2CKHA" "BCJJKN" "DKIM6U" "GDXWJ1" "TR5L57" "XC304E"
#>   [9] "1E8KD1" "5PW7WT" "AEP5EG" "BW10CL" "CFD12A" "CHJ9D2" "EMV4P6" "FTVE03"
#>  [17] "IRFJ09" "J3F6PD" "KXHGRH" "N4NV8B" "Q9LWGX" "RNQU14" "W5WIRP" "ZPS15A"
#>  [25] "09LFE4" "1CZM30" "2F6J3U" "3DTD2N" "3QHAFI" "55BPSE" "5XVTVH" "8IG767"
#>  [33] "9FRCIE" "ER464J" "FFGPS4" "FG6L7S" "GTLA8R" "P7RBPI" "S5H1GC" "TBCE78"
#>  [41] "YI16QD" "0HYZ23" "1VP3UC" "2F1IV1" "4LHK19" "59NYZE" "5EDLL7" "5IYDXN"
#>  [49] "6KLWVC" "80F2MI" "8TV4MT" "AZ4D19" "BTTHAJ" "CHSCFG" "DHNQ1W" "DRXMW4"
#>  [57] "EEGLWY" "FX9E4X" "G91ZM6" "I8ABC7" "J1R2EW" "LDND6J" "LN1DLY" "MQT080"
#>  [65] "NSIC4I" "PHB6TE" "PYPM1W" "QRWYQZ" "QW2Z3R" "RY1AZM" "W6MDVK" "WHQLH5"
#>  [73] "WQUN84" "XL658N" "XX0GYV" "YHHVC7" "YP910X" "Z904TJ" "ZQXZYB" "0IIAEN"
#>  [81] "0V4SAC" "0X1RZ9" "0XTZQ1" "465ERA" "55VDSQ" "5W621W" "653J82" "6MEP2C"
#>  [89] "76DIT4" "7ZNY75" "80KACX" "87AQLF" "967Y3D" "AIHJ8Z" "AZ3L0D" "CMMUKU"
#>  [97] "E3JP0C" "EX5K0S" "FLIZQI" "GM371F" "MEUZ85" "PA9F3J" "QCA36T" "S7IWWA"
#> [105] "SXSVEH" "T38W6H" "TJN1AD" "W0GUKI" "WJXIH9" "WNKKW3" "XZH41H" "3P9BX6"
#> [113] "7D09WH" "DGZLV3" "S63QDN" "ZW2X4N"
#>