obfuscatePed takes a pedigree object and creates aliases for all IDs and adjusts all date within a specified amount.

User provides a pedigree object (ped), the number of characters to be used for alias IDs (size), and the maximum number of days that the birthdate can be shifted (maxDelta).

obfuscatePed(
  ped,
  size = 6,
  maxDelta = 30,
  existingIds = character(0),
  map = FALSE
)

Arguments

ped

pedigree object

size

integer value indicating number of characters in alias IDs

maxDelta

integer value indicating maximum number of days that the birthdate can be shifted

existingIds

character vector of existing aliases to avoid duplication.

map

logical if TRUE a list object is returned with the new pedigree and a named character vector with the names being the original IDs and the values being the new alias values. Defaults to FALSE.

Value

An obfuscated pedigree

Examples

# \donttest{
library(nprcgenekeepr)
ped <- qcStudbook(nprcgenekeepr::pedGood)
obfuscatedPed <- obfuscatePed(ped)
ped
#>   id sire  dam sex gen      birth exit  age recordStatus
#> 1 d1 <NA> <NA>   F   0 2003-04-13 <NA> 19.0     original
#> 2 d2 <NA> <NA>   F   0 2002-06-22 <NA> 19.8     original
#> 3 s1 <NA> <NA>   M   0 2000-07-18 <NA> 21.8     original
#> 4 s2 <NA> <NA>   M   0 2005-06-19 <NA> 16.8     original
#> 5 o1   s1   d1   F   1 2015-02-04 <NA>  7.2     original
#> 6 o2   s1   d2   F   1 2009-03-17 <NA> 13.1     original
#> 7 o3   s2   d2   F   1 2012-04-11 <NA> 10.0     original
#> 8 o4   s2   d2   M   1 2008-04-13 <NA> 14.0     original
obfuscatedPed
#>       id   sire    dam sex gen      birth exit  age recordStatus
#> 1 76ZYBB   <NA>   <NA>   F   0 2003-03-20 <NA> 19.1     original
#> 2 VJ53TR   <NA>   <NA>   F   0 2002-06-23 <NA> 19.8     original
#> 3 E259K9   <NA>   <NA>   M   0 2000-08-06 <NA> 21.7     original
#> 4 3D59BL   <NA>   <NA>   M   0 2005-06-10 <NA> 16.9     original
#> 5 0Z0IPT E259K9 76ZYBB   F   1 2015-02-03 <NA>  7.2     original
#> 6 ADAUAK E259K9 VJ53TR   F   1 2009-02-27 <NA> 13.2     original
#> 7 98VPVU 3D59BL VJ53TR   F   1 2012-03-29 <NA> 10.1     original
#> 8 30IEZS 3D59BL VJ53TR   M   1 2008-05-03 <NA> 14.0     original
# }